requirements.txt
torch>=2.0.0
torchvision>=0.15.0
mmcv-full>=2.0.0
mmengine>=0.7.0
mmdet>=3.0.0
mmyolo>=0.6.0
opencv-python>=4.7.0
pillow>=9.0.0
pycocotools>=2.0.0
matplotlib>=3.5.0
numpy>=1.21.0
